header keywords update and ARmask update
